requirements.txt
fastapi
uvicorn
python-multipart
supabase
httpx
python-docx
pandas
openpyxl
python-dotenv
pydantic
tqdm
requests
aiofiles
python-dateutil
numpy
pymupdf
#faiss-cpu
torch
#--index-url https://download.pytorch.org/whl/cpu
transformers
sentence-transformers
huggingface-hub
llama-index-core
#llama-index-embeddings-huggingface
#llama-index-vector-stores-faiss
llama-index-llms-groq
#langchain
aiohttp
async-timeout==4.0.3
bleach
tenacity>=8.2.0
python-magic
